chore: unique EditToolbar, reuse components
