+++ src/resources/functions/edit.tsx
@@ -4,15 +4,12 @@ import CardContent from '@mui/material/CardContent';
 import deepEqual from 'deep-is';
 import { useEffect, useState } from 'react';
 import {
-    Button,
     EditBase,
     EditView,
     LoadingIndicator,
-    SaveButton,
     SelectInput,
     SimpleForm,
     TextInput,
-    Toolbar,
     useNotify,
     useRecordContext,
     useRedirect,
@@ -20,7 +17,6 @@ import {
     useTranslate,
 } from 'react-admin';
 import { useWatch } from 'react-hook-form';
-import { useNavigate } from 'react-router';
 import { FlatCard } from '../../components/FlatCard';
 import { FormLabel } from '../../components/FormLabel';
 import { EditPageTitle } from '../../components/PageTitle';
@@ -28,8 +24,8 @@ import { useSchemaProvider } from '../../provider/schemaProvider';
 import { FunctionIcon } from './icon';
 import { getFunctionUiSpec } from './types';
 import { JsonSchemaInput } from '../../components/JsonSchema';
-import ClearIcon from '@mui/icons-material/Clear';
 import { MetadataInput } from '../../components/MetadataInput';
+import { EditToolbar } from '../../components/toolbars/EditToolbar';
 
 const SpecInput = (props: {
     source: string;
@@ -79,25 +75,7 @@ const SpecInput = (props: {
         />
     );
 };
-export const FunctionEditToolbar = () => {
-    const translate = useTranslate();
-    const navigate = useNavigate();
-    const handleClick = () => {
-        navigate(-1);
-    };
-    return (
-        <Toolbar sx={{ justifyContent: 'space-between' }}>
-            <SaveButton />
-            <Button
-                color="info"
-                label={translate('actions.cancel')}
-                onClick={handleClick}
-            >
-                <ClearIcon />
-            </Button>
-        </Toolbar>
-    );
-};
+
 export const FunctionEdit = () => {
     const notify = useNotify();
     const redirect = useRedirect();
@@ -152,7 +130,7 @@ export const FunctionEdit = () => {
 
                     <EditView component={Box}>
                         <FlatCard sx={{ paddingBottom: '12px' }}>
-                            <SimpleForm toolbar={<FunctionEditToolbar />}>
+                            <SimpleForm toolbar={<EditToolbar />}>
                                 <FormLabel label="fields.base" />
 
                                 <Stack direction={'row'} spacing={3} pt={4}>
